Move genre.rb and music_album.rb files to classes folder
