+++ classes/music_album.rb
@@ -1,4 +1,4 @@
-require './item'
+require_relative '../item'
 
 class MusicAlbum < Item
   attr_reader :on_spotify, :archived
