+++ modules/display.rb
@@ -11,9 +11,10 @@ end
 
 def display_genres_on_list_all_genres(genres)
   genres.each do |genre|
-    puts "Name: #{genre['name']}" if genre.instance_of?(Hash)
-    puts "Name: #{genre.name}" if genre.instance_of?(Genre)
+    print "'#{genre['name']}', " if genre.instance_of?(Hash)
+    print "'#{genre.name}', " if genre.instance_of?(Genre)
   end
+  puts ''
 end
 
 def display_genres_on_add_music_album(genres)
