+++ spec/music_album_spec.rb
@@ -1,5 +1,5 @@
 require './item'
-require './music_album'
+require_relative '../classes/music_album'
 
 RSpec.describe MusicAlbum do
   describe '#initialize' do
